+++ extensions/podcast-style/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## 1.0.1 - 2026-09-14
+
+- Declared the extension permissions contract in the manifest.
+
 ## 1.0.0 - 2026-09-01
 
 - Initial release. Orange progress and a microphone icon on podcast tracks.